backend/requirements.txt
fastapi
uvicorn
python-dotenv

# Google APIs
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

# Resume parsing
pdfplumber
python-docx

# AI
groq

# Email (Gmail API)
google-auth-oauthlib
google-api-python-client
